+++ lib/chat.ex
@@ -1,9 +1,32 @@
 defmodule Chat do
   @moduledoc """
-  Chat keeps the contexts that define your domain
-  and business logic.
-
-  Contexts are also responsible for managing your data, regardless
-  if it comes from the database, an external API or others.
+  High level functions
   """
+
+  def db_get(key) do
+    case key do
+      {:file_chunk, file_key, first, last} -> read_chunk({first, last}, file_key)
+      _ -> Chat.Db.get(key)
+    end
+  end
+
+  def db_put(key, value) do
+    case key do
+      {:file_chunk, file_key, first, last} -> write_chunk(value, {file_key, first, last})
+      _ -> Chat.Db.put(key, value)
+    end
+  end
+
+  defp read_chunk(range, key) do
+    {data, _last} =
+      Chat.FileFs.read_exact_file_chunk(range, key, path())
+
+    data
+  end
+
+  defp write_chunk(value, chunk_params) do
+    Chat.FileFs.write_file(value, chunk_params, path())
+  end
+
+  defp path, do: CubDB.data_dir(Chat.Db.db()) <> "_files"
 end