+++ lib/chat/dialogs/private_message.ex
@@ -1,5 +1,7 @@
 defmodule Chat.Dialogs.PrivateMessage do
   @moduledoc "Represents decrypted message for a peer"
 
+  @type t :: %__MODULE__{}
+
   defstruct [:timestamp, :index, :type, :content, :is_mine?, :id]
 end