+++ lib/naive_api/schema.ex
@@ -5,6 +5,7 @@ defmodule NaiveApi.Schema do
   use Absinthe.Schema
 
   alias NaiveApi.Chat
+  alias NaiveApi.Data
   alias NaiveApi.Room
   alias NaiveApi.Upload
   alias NaiveApi.User
@@ -12,10 +13,6 @@ defmodule NaiveApi.Schema do
   import_types(NaiveApi.Schema.Types)
 
   query do
-    field :hello, :string do
-      resolve(fn _, _, _ -> {:ok, "Hello Wold"} end)
-    end
-
     @desc "List all users excluding requesting one"
     field :user_list, list_of(non_null(:card)) do
       arg(:my_public_key, :public_key |> non_null)
@@ -26,7 +23,7 @@ defmodule NaiveApi.Schema do
     Reads chat messages.
 
     `before` and `amount` are in indexes. One index contains at least one message.
-    If `before` omitted, most recent messaages will be returned.
+    If `before` omitted, most recent messages will be returned.
     If `amount` omitted, messages in 20 indexes will be returned.
     """
     field :chat_read, list_of(:message |> non_null) do
@@ -41,14 +38,29 @@ defmodule NaiveApi.Schema do
     Reads room messages.
 
     `before` and `amount` are in indexes. One index contains at least one message.
-    If `before` omitted, most recent messaages will be returned.
+    If `before` omitted, most recent messages will be returned.
     If `amount` omitted, messages in 20 indexes will be returned.
     """
-    field :room_read, list_of(:message |> non_null) do
+    field :room_read, list_of(:message |> non_null) |> non_null() do
       arg(:room_keypair, non_null(:input_key_pair))
       arg(:before, :integer)
       arg(:amount, :integer)
       resolve(&Room.read/3)
+    end
+
+    @desc """
+    Dumps all keys available
+    """
+    field :data_keys, list_of(:string |> non_null) do
+      resolve(&Data.all_keys/2)
+    end
+
+    @desc """
+    Gets value by key
+    """
+    field :data_value, :string do
+      arg(:key, :string |> non_null)
+      resolve(&Data.get_value/2)
     end
   end
 
@@ -56,7 +68,7 @@ defmodule NaiveApi.Schema do
     @desc """
     Registers new user in the system. Providing unique keypair.
 
-    There is no uniqeness restriction on the name.
+    There is no uniqueness restriction on the name.
     The system references users by their public_key 
     """
     field :user_sign_up, non_null(:identity) do